Added a TracerSlicer, which allows one to perform multiplane ray tracing with planes at fixed redshift inervals which are different from the redshifts of the galaxies. The galaxies are paired to their closest plane in redshift space.
This lead to a large refactor of the plane / ray-tracing modules,
which involed allowing planes to exist which contain no galaxies (in
case a sliced plane in the tracer above happens to have no nearby
galaxies).

Also removed 'stack' modules and will implement multi-image lensing in a
more elegent way.